shelves/requirements.txt
flask-bootstrap
flask-nav
flask-misaka
flask-uploads
flask-mysqldb
